SQL Server sync version of the expected files are added. Tests are all passing. Started working on the Postgres side of things. Fixed the swapped mistake for the file naming of the repositories. Added new debug method for Postgres.
