+++ SpotWelder.DatabaseForTests/Postgres/create database spot_welder.sql
@@ -1,0 +1,10 @@
+CREATE DATABASE spot_welder
+    WITH
+    OWNER = postgres
+    ENCODING = 'UTF8'
+    LC_COLLATE = 'en_US.utf8'
+    LC_CTYPE = 'en_US.utf8'
+    LOCALE_PROVIDER = 'libc'
+    TABLESPACE = pg_default
+    CONNECTION LIMIT = -1
+    IS_TEMPLATE = False;
