+++ SpotWelder.DatabaseForTests/Postgres/public/Tables/data_type_test.sql
@@ -1,0 +1,27 @@
+﻿CREATE TABLE public.data_type_test (
+	data_type_test_id int4 GENERATED ALWAYS AS IDENTITY( INCREMENT BY 1 MINVALUE 1 MAXVALUE 2147483647 START 1 CACHE 1 NO CYCLE) NOT NULL,
+	int64 int8 NOT NULL,
+	bytearray_binary BYTEA CHECK (octet_length(bytearray_binary) = 10) NOT NULL,
+	boolean bool NOT NULL,
+	string_char char(1) NOT NULL,
+	datetime_date date NOT NULL,
+	datetime_datetime timestamp(3) NOT NULL,
+	datetime_datetime2 timestamp(6) NOT NULL,
+	datetimeoffset timestamptz NOT NULL,
+	decimal_decimal decimal(4,2) NOT NULL,
+	double double precision NOT NULL,
+	int32 int4 NOT NULL,
+	decimal_money money NOT NULL,
+	string_nchar char(1) NOT NULL,
+	decimal_numeric numeric(4,2) NOT NULL,
+	string_nvarchar varchar(10) NOT NULL,
+	single real NOT NULL,
+	datetime_smalldatetime timestamp(0) NOT NULL,
+	int16 smallint NOT NULL,
+	timespan time NOT NULL,
+	guid uuid NOT NULL,
+	bytearray_varbinary BYTEA CHECK (octet_length(bytearray_varbinary) <= 10) NOT NULL,
+	string varchar(10) NOT NULL,
+	xml xml NOT NULL,
+	CONSTRAINT pk_data_type_test_id PRIMARY KEY (data_type_test_id)
+)
