+++ SpotWelder.DatabaseForTests/Postgres/public/Tables/data_type_test_nullable.sql
@@ -1,0 +1,27 @@
+﻿CREATE TABLE public.data_type_test_nullable (
+	data_type_test_nullable_id int4 GENERATED ALWAYS AS IDENTITY( INCREMENT BY 1 MINVALUE 1 MAXVALUE 2147483647 START 1 CACHE 1 NO CYCLE) NOT NULL,
+	int64 int8 NULL,
+	bytearray_binary BYTEA CHECK (octet_length(bytearray_binary) = 10) NULL,
+	boolean bool NULL,
+	string_char char(1) NULL,
+	datetime_date date NULL,
+	datetime_datetime timestamp(3) NULL,
+	datetime_datetime2 timestamp(6) NULL,
+	datetimeoffset timestamptz NULL,
+	decimal_decimal decimal(4,2) NULL,
+	double double precision NULL,
+	int32 int4 NULL,
+	decimal_money money NULL,
+	string_nchar char(1) NULL,
+	decimal_numeric numeric(4,2) NULL,
+	string_nvarchar varchar(10) NULL,
+	single real NULL,
+	datetime_smalldatetime timestamp(0) NULL,
+	int16 smallint NULL,
+	timespan time NULL,
+	guid uuid NULL,
+	bytearray_varbinary BYTEA CHECK (octet_length(bytearray_varbinary) <= 10) NULL,
+	string varchar(10) NULL,
+	xml xml NULL,
+	CONSTRAINT pk_data_type_test_nullable_id PRIMARY KEY (data_type_test_nullable_id)
+)
