+++ SpotWelder.IntegrationTests/ModelGenerationTests/ExpectedModels/SerializeCsv.sync.sqlserver.expected
@@ -1,0 +1,50 @@
+//https://www.nuget.org/packages/CsvHelper/
+using CsvHelper;
+using CsvHelper.Configuration;
+using System.Globalization;
+
+namespace Namespace1
+{
+    public partial class SerializationService
+    {
+        private CsvConfiguration GetCsvConfig()
+        {
+            var config = new CsvConfiguration(CultureInfo.CurrentCulture)
+            {
+                Delimiter = "|",
+                HasHeaderRecord = false
+            };
+
+            return config;
+        }
+
+        public string ToCsv(IEnumerable<DataTypeTestEntity> input)
+        {
+            using (var writer = new StringWriter())
+            {
+                using (var csvWriter = new CsvWriter(writer, GetCsvConfig()))
+                {
+                    csvWriter.WriteRecords(input);
+                }
+
+                return writer.ToString();
+            }
+        }
+
+        public IEnumerable<DataTypeTestEntity> FromCsv(string csv)
+        {
+            using (var reader = new StringReader(csv))
+            {
+                using (var csvReader = new CsvReader(reader, GetCsvConfig()))
+                {
+                    var records = csvReader.GetRecords<DataTypeTestEntity>();
+
+                    //Has to be forced to a List otherwise a NullReference Exception will occur
+                    var lst = records.ToList();
+
+                    return lst;
+                }
+            }
+        }
+    }
+}
