+++ SpotWelder.IntegrationTests/ModelGenerationTests/ExpectedModels/SerializeJson.sync.sqlserver.expected
@@ -1,0 +1,12 @@
+//https://www.nuget.org/packages/Newtonsoft.Json/
+using Newtonsoft.Json;
+
+namespace Namespace1
+{
+    public partial class SerializationService
+    {
+        public string ToJson(IEnumerable<DataTypeTestEntity> input) => JsonConvert.SerializeObject(input);
+
+        public IEnumerable<DataTypeTestEntity> FromJson(string json) => JsonConvert.DeserializeObject<IEnumerable<DataTypeTestEntity>>(json);
+    }
+}
