+++ SpotWelder.IntegrationTests/ModelGenerationTests/ExpectedModels/Service.sync.sqlserver.expected
@@ -1,0 +1,50 @@
+using System;
+
+namespace Namespace1
+{
+    public class DataTypeTestService
+    {
+        private readonly IDataTypeTestRepository _repository;
+        private readonly IDataTypeTestValidation _validation;
+
+        public DataTypeTestService(
+            IDataTypeTestRepository repository,
+            IDataTypeTestValidation validation)
+        {
+            _repository = repository;
+            _validation = validation;
+        }
+
+        public DataTypeTestEntity? GetDataTypeTest(int dataTypeTestId)
+        {
+            Validations.IsGreaterThanZero(dataTypeTestId, nameof(dataTypeTestId));
+
+            var dbEntity = _repository.Using(x => x.Select(dataTypeTestId));
+
+            return dbEntity;
+        }
+
+        public DataTypeTestEntity Add(DataTypeTestEntity? entity)
+        {
+            Validations.IsValid(_validation, entity, nameof(entity));
+
+            entity.DataTypeTestId = _repository.Using(x => x.Insert(entity));
+
+            return entity;
+        }
+
+        public void Edit(DataTypeTestEntity entity)
+        {
+            Validations.IsNotNull(entity, nameof(entity));
+
+            _repository.Using(x => x.Update(entity));
+        }
+
+        public void Remove(int dataTypeTestId)
+        {
+            Validations.IsGreaterThanZero(dataTypeTestId, nameof(dataTypeTestId));
+
+            _repository.Using(x => x.Delete(dataTypeTestId));
+        }
+    }
+}
