+++ SpotWelder.Lib/Services/Generators/RepositoryDapperGenerator.cs
@@ -75,7 +75,7 @@ namespace SpotWelder.Lib.Services.Generators
 			template.Replace("{{DynamicParametersInsert}}", FormatDynamicParameterList(lstInsert));
 			template.Replace("{{DynamicParametersUpdate}}", FormatDynamicParameterList(instructions.Properties));
 			
-			var rt = instructions.Elections.HasFlag(GenerationElections.RepoStatic) ? "Static" : string.Empty;
+			var rt = instructions.Elections.HasFlag(GenerationElections.RepoStatic) ? "Dapper" : string.Empty;
 			
       return GetFormattedCSharpResult($"{instructions.ClassName}{rt}Repository.cs", template);
 		}
