+++ SpotWelder.Ui/Profile.json
@@ -4,6 +4,26 @@
     "ConnectionStrings": [
       {
         "SqlEngine": 0,
+        "ConnectionString": "Data Source=.;Initial Catalog=SpotWelder;Integrated Security=True;Encrypt=False",
+        "Verified": true
+      },
+      {
+        "SqlEngine": 1,
+        "ConnectionString": "Host=localhost;Database=millions_of_things;Username=postgres;Password=postgres",
+        "Verified": true
+      },
+      {
+        "SqlEngine": 0,
+        "ConnectionString": "Data Source=.;Initial Catalog=ScratchSpace;Integrated Security=SSPI;",
+        "Verified": true
+      },
+      {
+        "SqlEngine": 1,
+        "ConnectionString": "Host=localhost;Database=spot_welder;Username=postgres;Password=postgres",
+        "Verified": true
+      },
+      {
+        "SqlEngine": 0,
         "ConnectionString": "Data Source=whatever;Initial Catalog=;Integrated Security=False;User ID=;Password=;Encrypt=False",
         "Verified": true
       },
@@ -22,19 +42,10 @@
         "ConnectionString": "Host=12345;Database=qwertty;Username=userguy;Password=secretpassword;Port=6553",
         "Verified": true
       },
-      {
-        "SqlEngine": 1,
-        "ConnectionString": "Host=localhost;Database=millions_of_things;Username=postgres;Password=postgres",
-        "Verified": true
-      },
+
       {
         "SqlEngine": 1,
         "ConnectionString": "Host=asdfasd;Database=asdfasdfasdf;Username=qwerqwer;Password=",
-        "Verified": true
-      },
-      {
-        "SqlEngine": 0,
-        "ConnectionString": "Data Source=.;Initial Catalog=ScratchSpace;Integrated Security=SSPI;",
         "Verified": true
       },
       {
